+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added:
+- Automatic redirect from localhost:3000 to localhost:3000/browse/ for user convenience.
+
 ## [0.7.0] - 2025-10-08
 
 ### Added: